+++ Sources/MapboxSearch/PublicAPI/Offline/SearchCancelable.swift
@@ -2,11 +2,11 @@ import Foundation
 
 /// A type that conforms to `Cancelable` typically represents a long
 /// running operation that can be canceled.
-public protocol Cancelable: AnyObject {
+public protocol SearchCancelable: AnyObject {
     func cancel()
 }
 
-internal final class CommonCancelableWrapper: Cancelable {
+internal final class CommonCancelableWrapper: SearchCancelable {
     private let cancelable: MapboxCommon.Cancelable
 
     internal init(_ cancelable: MapboxCommon.Cancelable) {