+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 ### Fixed
 
  - Suggestion resolve no longer cancel on search query change.
+ - Cancelable renamed into SearchCancelable to prevent naming conflicts with Maps SDK.
 
 ## [1.0.0-beta.21] - 2021-12-22